Remove task.h and task-impl.h
