+++ core/traffic_class_test.cc
@@ -8,8 +8,6 @@
 
 #include "module.h"
 #include "scheduler.h"
-#include "task.h"
-#include "task-impl.h"
 #include "traffic_class.h"
 
 #define CT TrafficClassBuilder::CreateTree
